refactor(services): split cell/combat.rs into submodules
Separate QR damage pipeline, dead-state flags, and NPC threat
generation into damage.rs/state.rs/threat.rs. mod.rs re-exports
all prior public items.

Public API unchanged. No behavior changes.
